refactor to use existing softmax function and move parallel_tv_prox to optimize.py
